# Compute exact `log` when base is a power of 2 (also faster)