Create a `my_log2` C function, that deals with logs of exact powers of 2. This will give us exact results when the number is exact and the base is a small power of two:

```scheme
(log (expt 8 21) 8) => 21
(log (expt 2 10) 2) => 10
(log 1/4 2)         => -2
(log -4 2)          => 2+4.53236014182719i
```

In the last example, we got an _exact_ real part.

Roughly,
* For fixnums, compare to the known possible powers of two (there are not too many of them anyway)
* For bignums, use `mpz_sizeinbase`
* For rationals, do $\log(x/y) = \log(x) - \log(y)$. This allows us to give exact answers to `(log 1/64 4)`.

But additionally,

* If the base itself is a small power of two, we try to get an exact result anyway: use it in `mpz_sizeinbase` for bignums and use it (the base) to check for known powers for fixnums
* If the number is negative, return a complex with an imaginary part equal to $\pi/\log(base)$

This seems faster than the previous code, in particular when the first argument is a fixnum (but there seems to also be some speedup for bignums). Some timings:

```scheme
(define a  #f)

(define y (expt 4 50))
(time (repeat 50000000 (set! a (log y 4))))
;; current code  7019.777 ms
;; new code      4276.111 ms

(define y (expt 16 100))
(time (repeat 50000000 (set! a (log y 16))))
;; current code 6961.37 ms
;; new code     4384.379 ms

(define y (expt 2 500))
(time (repeat 50000000 (set! a (log y 2))))
;; current code     6929.235 ms
;; new code         4639.744 ms

(define x (expt 4 20))
(time (repeat 50000000 (set! a (log x 4))))
;; current code 6385.002 ms
;; new code     1645.658 ms

(define x (expt 2 50))
(time (repeat 50000000 (set! a (log x 2))))
;; current code 6365.162 ms
;; new code     2401.009 ms
```