@jpellegrini 2023-11-27
It would be possible to extend this to bases up to `10`, for example, but it would require some more computation at the beginning of `my_log2`. I'm not sure it's worth it. Gambit, for example, only gives exact answers when the base is a power of 2, and returns inexact answers for all other bases. All other schemes seem to not preserve exactness for `log`. So if we keep this PR as it is, we'd have something similar to Gambit (and faster than what STklos has today).