@jpellegrini 2023-11-27
> Not an error, but it would be nice in the future to enhance this to get `1/2` there.

Just did it. :)

Now the only thing missing is when the result is an exact rational with numerator different from one:

```scheme
(log 8 4)  => 1.5
```

I'd like it to answer `3/2`, but that requires more work. Maybe in a future PR...