@jpellegrini 2023-11-27
> I'd like it to answer `3/2`, but that requires more work. Maybe in a future PR...

Okay... So just because I said that, I immediately realized it's actually a two-line change. implemented, and now it's ready. :)

```scheme
    (log 4 8) => 2/3
    (log 8 4) => 3/2
```