@egallesio 2023-12-04
Hi @jpellegrini,

That's really really great. I'll integrate, it asap. 
Thanks.